Modularize basic modules checkpointing
